microservices/old-finder/requirements.txt
Flask
flask_restful
mongoengine==0.23.1
flask-mongoengine
flask_mongorest
mimerender
cleancat
python-dotenv
Flask-WTF
wheel
bcrypt
uuid
redis
pycurl
requests
awesome-slugify
datetime
pymongo